+++ gqlgen.yml
@@ -18,6 +18,8 @@ models:
     model: github.com/icco/graphql.Geo
   Link:
     model: github.com/icco/graphql.Link
+  Log:
+    model: github.com/icco/graphql.Log
   Post:
     model: github.com/icco/graphql.Post
   Tweet:
